Move validateVariantAgainstQuestion to variant model code
