+++ apps/prairielearn/src/models/variant.ts
@@ -1,6 +1,8 @@
+import { HttpStatusError } from '@prairielearn/error';
 import { loadSqlEquiv, queryAsync, queryOptionalRow, queryRows } from '@prairielearn/postgres';
 
 import { IdSchema, SubmissionSchema, Variant, VariantSchema } from '../lib/db-types.js';
+import { idsEqual } from '../lib/id.js';
 
 const sql = loadSqlEquiv(import.meta.url);
 
@@ -55,4 +57,27 @@ export async function selectVariantsByInstanceQuestion({
       max_submission_score: SubmissionSchema.shape.score.unwrap(),
     }),
   );
+}
+export async function validateVariantAgainstQuestion(
+  unsafe_variant_id: string,
+  question_id: string,
+  instance_question_id: string | null = null,
+): Promise<Variant> {
+  const variant = await selectVariantById(unsafe_variant_id);
+  if (variant == null || !idsEqual(variant.question_id, question_id)) {
+    throw new HttpStatusError(
+      400,
+      `Client-provided variant ID ${unsafe_variant_id} is not valid for question ID ${question_id}.`,
+    );
+  }
+  if (
+    instance_question_id != null &&
+    (!variant.instance_question_id || !idsEqual(variant.instance_question_id, instance_question_id))
+  ) {
+    throw new HttpStatusError(
+      400,
+      `Client-provided variant ID ${unsafe_variant_id} is not valid for instance question ID ${instance_question_id}.`,
+    );
+  }
+  return variant;
 }
